refactor: rename module to websearch.py
